Split links into their own files
