+++ packages/peregrine/lib/Apollo/links/mutationQueueLink.js
@@ -1,0 +1,5 @@
+import MutationQueueLink from '@adobe/apollo-link-mutation-queue';
+
+export default function createMutationQueueLink() {
+    return new MutationQueueLink();
+}
